hpcprof: Fix crashes related to fs::canonical
A recent change to hpcprof uses `fs::canonical` at various points to
create absolute paths with symlinks resolved. Unfortunately errors are
not properly handled in these cases, `fs::canonical` throws an exception
if the file does not exist, causing errors or crashes. In particular
an empty path is common, e.g. when passing `-S`.

Switch this to `fs::weakly_canonical`, which doesn't throw an exception
and instead performs a best-effort match.