Remove dependence on Boost.Graph internals
hpcstruct uses internal functions of Boost.Graph to parse the DOT files
nvdisasm generates. This was presumably to parse port information in the
graph which is used to represent "superblock" arcs, but after rewrites
in other locations this information is no longer used in practice.

Replace the DOT parsing code with the Boost.Graph public interface, and
convert into the prior data structures with minimal changes.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/815

Signed-off-by: Jonathon Anderson <janderson@rice.edu>