hpcstruct: Fill in gaps in CUDA bins before Volta
Before Volta CUDA binaries have 8-byte instruction words, except every
triplet of instructions is preceded by an 8-byte (control?) word. This
word is not documented as an instruction by nvdisasm and cuobjdump,
which leads to odd gaps and incorrect loop attribution in the final
Structfile.

Insert a no-op instruction where these gaps are expected as part of the
basic block of the following instruction. The end result is that these
control(?) words are now properly attributed to loop nests, instead of
only having line numbers but no outer structure.

Signed-off-by: Jonathon Anderson <janderson@rice.edu>