hpcstruct: Fix CUDA SASS parsing for VLIW insts
The CUDA SASS output contains VLIW instructions, written as two normal
instructions on two separate lines surrounded by braces. The parser in
hpcstruct removes the first two bytes of the latter line, but this
incorrectly removes part of a port label (<...>) or the address itself.
Very clearly a bug, but present since the very first implementation of
this parser.

Remove this modification, which fixes cases where the offset failed to
parse later in the function and caused entire sub-CFGs to be missing in
the final Structfile output.

Signed-off-by: Jonathon Anderson <janderson@rice.edu>