ci: Update ci.predeps to 1.x
The main change in the full release (aside from a few syntax tweaks) is
the addition of more retry logic, which should help keep the CI from
needing a restart as much as it does today.

With the full release the component is now part of the CI/CD catalog,
which means it will now update automatically until the next major
version bump.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>