hpcstruct: Remove hacky control flow at CUDA function end
Some extra padding blocks are added at the end of CUDA functions so that
the total size in the CFG matches the total size of the function symbol.
These blocks are connected to the CFG via a DIRECT edge from the last
instruction in the function. This results in an incorrectly crafted
CFG if the last instruction happens to be an unconditional jump.

As it happens, the padding blocks do not add any new information to the
Structfile, so remove them to produce a consistent and correct CFG.
In some cases the line number changes as a result, vanishing as reported
in [#845].

[#845]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/845

Signed-off-by: Jonathon Anderson <janderson@rice.edu>