hpcstruct: Stop writing out the nvidia/ directory
Currently hpcstruct always writes out an nvidia/ directory beside the
output Structfile, which contains DOT files for all the CUDA binaries
processed. These files are never read back in and are really an
implementation detail, so there's no point in writing them in the first
place. It also is a duplicate call to `nvdisasm` which can't possibly
be very fast.

Stop writing these files out and restrict the output to just the
Structfile itself.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>