hpcstruct: XFAIL CUDA tests on Elfutils >=0.191
Bug [#831] affects all `hpcstruct` output for CUDA binaries, regardless
of the value of `--gpucfg`. Previously the XFAIL was only added for the
case where `--gpucfg=yes`, extend it to include `--gpucfg=no` as well.

This fixes the failures in the `spack +cuda` job, which were combining
CUDA with the latest Elfutils.

[#831]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/831