Update CI to latest CUDA versions
Update the CI to test CUDA versions that will be supported for the rest
of the year, specifically CUDA 12.x (12.0 to 12.5). CUDA 11.x is no
longer tested in CI (except as part of an Ubuntu repackaging).

Technically, the CUDA Toolkit doesn't have a clear EOL policy. But the
driver it ships with does, and that date is used as [the EOL of the
corresponding container images][1]. The versions used here are based on
the EOL of CUDA 11.x on July 1st, as well as catching up on the releases
of CUDA 12.4 and 12.5.

[1]: https://gitlab.com/nvidia/container-images/cuda/-/blob/master/doc/support-policy.md?ref_type=heads

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>