tests: XFAIL failures caused by bug in CUDA 12.3
Prior to CUDA 12.4, nvdisasm does not generate control flow edges for
the `CALL.REL` set of instructions, which are otherwise identical to
an unconditional jump (`BRA`) and don't represent a call to another
function symbol.

Mark any `hpcstruct` tests on binaries containing this instruction as
an XFAIL when running on an CUDA from before this change.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>