ci: Add autoformatting with clang-format
We don't currently have a code style, but having an autoformatter is a
significant boon to productivity when dealing with large code segments
or integrating code from multiple sources. This commit adds the basic
skeleton for autoformatting: a .clang-format stub file and enforcement
CI job.

Subdirectories can add their own formatting configurations to override
the stub, but to avoid excessive bikeshedding the only directory that
does this right now is `tests/`.