ci: Lint C/C++ code with clang-tidy
Clang-tidy is a linter/static analyzer for C/C++ from the LLVM project.
While we don't require that every issue presented is fixed before merge,
we do want to report the issues as an aid to review. This commit adds a
CI job to run clang-tidy and report the issues it finds to GitLab.

The checks enabled are catch-all generic (and only lightly opinionated)
checks for portability, performance and good code practice.