Update wraps to the WrapDB
The libpfm and bzip2 wraps has been uploaded to the shared Meson WrapDB,
a public MIT-licensed project that hosts wraps for projects that don't
already have Meson build support. This contributes the wraps back to the
Meson users community, and is that much more permanent than the fork
used to host them today.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>