Convert OneTBB wrap to method = cmake
OneTBB is built with CMake and has no dependencies, meaning it should
really be using a `method = cmake` wrap instead of a ported Meson build
script. Convert it to a `method = cmake` wrap homed here, like the other
CMake-powered wraps.

The only reason we used a Meson build port was because of a Meson bug
related to how link flags from CMake subprojects are handled, see [1]
for details. This should be fixed in the to-be-released Meson 1.6.0, but
for earlier versions an extra layer of indirection is used to inject a
partial_dependency() method call and strip the problematic flags.

[1]: https://github.com/mesonbuild/meson/pull/13422

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>