Absorb the libiberty wrap
Libiberty is a fragment of binutils/gdb/etc. that provides a number of
fallbacks for functions usually implemented in Glibc. Using it comes
with a number of caveats and there are strong recommendations to [avoid
using libiberty in new code][1]. Despite this, both we and Dyninst use
the demangler implemented in libiberty even though other demangler
interfaces exist that don't have these caveats (see [2]).

In light of this, the wrap for libiberty is not suitable for
contribution back to the WrapDB, so it is re-homed to this repository
as an inline wrap still licensed under MIT. As a slight benefit, the
wrap is adjusted to use binutils as the source for libiberty instead of
gdb (which its source in the Debian/Ubuntu packaging).

[1]: https://packages.debian.org/sid/libiberty-dev
[2]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/828

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>