Remove internal HPCRUN_AUDIT_MAIN_LIB env variable
Currently, the hpcrun auditor uses the HPCRUN_AUDIT_MAIN_LIB internal
environment variable to identify the "mainlib" (libhpcrun). This works,
but causes a segfault if the variable is unset, not set correctly or
otherwise doesn't exactly match the path for libhpcrun.so.

Support cases where the exact path to libhpcrun.so isn't known
beforehand by matching using the "libhpcrun.so" basename instead.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>