hpcrun: Prepare fnbounds for integration
This commit slashes down hpcfnbounds to the bare essentials that can be
integrated into hpcrun. Global variables are turned into function
arguments, dead values and headers are removed, any functionality
outside of the "server" use case (including debugging) is deleted,
functions are given `fnb_` prefixes for namespacing, etc.

The end result is that the core functionality that fnbounds provides
is exposed as a single function call, and the server bits have no logic
outside of communicating the bits back to hpcrun.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>