Add an ubercontainer as the main devcontainer
Until now, the devcontainers have directly reflected the environments
tested in CI. Each environment is unique and debugging problems that
aren't easily reflected in any one place requires jumping between a
number of containers and building in all of them. This is a slow and
inefficient use of precious developer time, and is by no means
understandable to the casual contributor. We need a container designed
from the ground up for development.

This commit adds the main devcontainer, listed as the first and default
devcontainer and thus easy to recommend for development. This container
is an "ubercontainer," a container that contains all the software
HPCToolkit could possibly use all in one container.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>