ci: Use the uber-devcontainer for clang-tidy
There are a number of files for which the clang-tidy check in CI
produces compile errors. Some (but not all) of these are because some
dependencies are not present in the ubuntu24.04 container where the
check runs, not all files are expected to compile without e.g. CUDA.

The uber-devcontainer has all these dependencies installed. Swap the
clang-tidy check to use this container to resolve the most clang-tidy
errors that are caused by missing dependencies.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>