Add a basic Gitpod configuration
[Gitpod] is one (of many) services that provide cloud/VM-backed
development environments. The tagline is to develop from "anywhere"
straight from the browser, but (unlike a pure web editor) still allow
compiling and running code. Gitpod happens to have an integration with
GitLab [1], if enabled it allows a one-click path to spin up a Gitpod
workspace from a branch or MR of the repository.

This commit configures Gitpod to use the uber-devcontainer as the base
container for the workspaces created from this branch. The end result is
an easy path to spin up a rich development environment from "anywhere."

[Gitpod]: https://gitpod.io
[1]: https://docs.gitlab.com/ee/integration/gitpod.html

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>