Convert some "unit tests" to GoogleTest
There are some source files that contain a short "unit test" at the end
of the file. These have have not been built or run in CI (and likely not
since their initial writing), making them less tests and more historical
artifacts.

Most of these "tests" just print a whole bunch of output, unfortunately
the "expected output" was never recorded or even documented. Still, some
are straightforward enough that an educated guess can be made.

This commit converts a selection of these tests to GoogleTest, complete
with the expectations/assertions of proper behavior.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>