ci: Improve determination of external pipelines
There are some CI jobs that require special hardware unavailable except
in the HPCToolkit GitLab group. Currently these jobs are disabled in
all MRs from forks, which tends to disable them when unnecessary.

Simplify the check to just "is this pipeline in the same project as the
MR target." Or something like that.

Signed-off-by: Jonathon Anderson <janderson@rice.edu>