clang-tidy: Disable some "unsafe" API checks
The `clang-analyzer-security.insecureAPI.DeprecatedOrUnsafeBufferHandling`
check warns on the use of a number of interfaces that gained
"bounds-checked" variants as part of the C11 standard (Annex K). These
variant interfaces have been the source of much argument and most are
not available in any version of Glibc.

Disabling the check removes ~500 or so false positives.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>