hpcrun: Refactor the attach protocol
Currently the auditor detects when libhpcrun.so is loaded and calls
`hpcrun_auditor_attach`. This runs into errors if libhpcrun.so is
dlopen'd due to an undocumented "feature" of Glibc where the LD_AUDIT
callback occurs before relocations are processed.

This commit reverses this call (libhpcrun.so calls
`hpcrun_connect_to_auditor`) so that the attachment works regardless of
when or how libhpcrun.so is loaded. This symbol is fetched via a
`dlsym()` to avoid a dependency on lazy evaluation of undefined symbols.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>