hpcrun: Disallow undefined symbols during the link
The new foils interface now carefully manages many of the externally
visible symbols, there should no longer be a scenario where libhpcrun.so
needs to call a symbol that will be "defined later."
Remove the exception and build everything with `-Wl,--no-undefined`.

The last loose symbol is an interaction between the Intel and PAPI
measurement subsystems. It is dead code and is safely removed.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/796

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>