mark stack in tramp.S as non-executable
The ELF linker now generates a warning message if the output binary contains a
segment with all three of the read, write and execute permission bits
set. These warnings are intended to help developers identify programs
which might be vulnerable to attack via these executable memory regions.

Adding the line ".section .note.GNU-stack,"",@progbits" to the assembly
code tells the linker to not use executable stack.

Signed-off-by: John Mellor-Crummey <johnmc@rice.edu>