force level zero immediate queue to be synchronous
when a level zero immediate queue is asynchronous, hpcrun
may inspect operation events for timestamps before the operations
are complete. for now, force the execution of these operations to
be synchronous so that hpcrun can obtain event timings.

FIXME: figure out how to properly extract timings for events
when a level zero immediate queue is asynchronous.

Signed-off-by: John Mellor-Crummey <johnmc@rice.edu>