Make Clang-Format style configuration global
Eventually we will enforce a consistent code style across the codebase.
Currently we don't, but subdirectories can opt-in to code formatting by
overriding the top-level .clang-format configuration with a code style
to use. This is liable to result in inconsistent styles for different
subdirectories, which isn't what we want.

In addition, it's an opt-in rather than an opt-out, which means new
subdirectories if they appear will be left unformatted. This only
aggravates the current problem.

This commit rearranges the .clang-format configuration, now the
top-level .clang-format file contains the style configuration and is
*enabled by default*. Subdirectories that do not want to be formatted
now opt-out of the formatting with their own .clang-format (which can
then be opted-in again but a deep subdirectory).

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>