Realign global .clang-format to LLVM's style
We directly refer to the LLVM code style in CONTRIBUTING.md, but the
style we use today is actually a derivative of the Google code style.
The overridden options are also not given any rationale for why the
default from the base code style are less preferred.

This commit changes the code style defined in .clang-format to be a
derivative of the LLVM code style. The deviations are commented with
rationale explaining why the change is made and occasional examples.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>