hpcrun: Refactor the foils subsystem
The foils subsystem covers some aspects of symbol interception, but in
preparation for larger changes we really need it to cover all aspects in
which libhpcrun.so calls in or out of the application namespace.
This commit refactors the foils to improve this coverage.

The function pointers used by the libmonitor, CUDA, ROCm, Level Zero,
GTPin, OpenCL, and PAPI-CUDA measurement subsystems are replaced by
dispatch-tables and/or wrapper functions implemented by the refactored
foils subsystem. This naturally keeps such implementation details out of
the already complex measurement subsystem codes.

The fake auditor is also a foil, but currently it directly implements
the audit ABI instead of following the more general foil architecture.
This grandfathering is not expected to last long, a future refactor to
the audit ABI will align it more with foils and remove this deviation.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>