devcontainers: Also enable CAP_PERFMON
In some cases, having CAP_SYS_ADMIN is not enough to enable all of the
perf_events interface, sometimes CAP_PERFMON must *also* be present.
It's not clear whether this is a quirk in Linux 6.10.9 or Podman's
default seccomp settings, but we can easily provide both capabilities so
just do so.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>