hpcprof: Hotfix #853
If a source file for a binary is recorded but has since been deleted,
hpcprof will still attempt to copy the file but will fail. In most cases
this scenario is silently ignored, but in this case specifically it will
noisily generate a WARNING.

There are multiple issues here, including technical debt and the
question of whether we *should* WARN here, but for now ensure the path
exists before passing it on to avoid this issue.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/853

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>