ci: Use stock Boost in Ubuntu 22.04
Previously we needed a custom build of Boost with `visibility=default`
to use CUDA support, so the Ubuntu 22.04 image built a custom Boost with
this option set. This was fixed (see [#815]), so this custom build is
no longer necessary for correct testing.

Remove the custom build and use the stock Boost shipped by Ubuntu.

[#815]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/815

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>