Use explicit allowlist for Clang-Formatting
Currently Clang-Formatting is disabled in most subdirectories using
local `.clang-format` files. While this works, it only operates at the
granularity of individual directory trees which makes it difficult to
opt-in to autoformatting for smaller changes at the granularity below a
subdirectory.

The Meson `clang-format` target supports specifying the files to
autoformat as an allowlist (`.clang-format-include`) plus denylist
(`.clang-format-ignore`). This commit uses these lists instead of style
configurations to restrict the set of autoformatted files, making it
possible to opt-in to autoformatting at the granularity of individual
source files.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>