Refine dependencies only needed for tests
Some dependencies are only necessary for building and running tests.
Previously these "extra" dependencies were placed under the feature flag
`-Dextended_tests` because *some* tests could be built without the extra
dependencies. This is confusing and liable to mysteriously missing
tests, and the set of no-dependency tests is shrinking with the
introduction of GoogleTest.

This commit renames the feature flag to just `-Dtests`, and it now
defaults to `enabled` instead of `auto`. This means that by default a
development build will have all possible tests and fail `meson setup` if
any dependencies needed by tests are missing.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>