ci: Update Spack (pre-)deployment checks
The current Spack deployment checks are failing in `develop`. This is a
good opportunity to update the checks, to make use of new Spack features
and bugfixes and to simplify the process as much as possible. As well as
extend the matrix of tested configurations to include more variants.

The `spack*` jobs have been moved into a sub-pipeline under ci/spack/,
which is now simply a Spack environment configured by environment
variables. Setting up mirrors and buildcaches is now handled entirely
by the CI script, and the set of mirrors is now near-entirely constant.

The functional change is that these jobs are now manual-only when in an
MR pipeline. This avoids spending an excess of a contributor's CI time
while still allowing maintainers/reviewers/the contributor themself to
run the tests for inspection before merge.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>