Update fallbacks for badly written modulefiles
For many Intel libraries, we implemented a number of fallbacks to
`find_library` to support badly designed modulefiles where CMake and/or
pkg-config was not properly configured to access the libraries.
These fallbacks were added out of necessity for working on this system,
we have since pushed for these modules to be fixed. So here we drop the
fallbacks and simplify the dependencies to the "normal" scenarios.

On the other hand, OpenCL is broken, and with the new defaults for
test dependencies it now causes a default build to fail. So introduce a
`find_library()` fallback for that one, and hope we can remove it soon.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>